+++ src/langs/php.mjs
@@ -82,7 +82,7 @@ export default {
 	},
 	string_single: {
 		end: {
-			regex: /"/g,
+			regex: /'/g,
 			switch_state: "php"
 		}
 	}