+++ src/langs/rust.mjs
@@ -33,8 +33,7 @@ export default {
 			regex: /\/\*/g,
 			switch_state: "comment_multi"
 		},
-		string_double: { regex: /"/g, switch_state: "string_double" },
-		string_single: { regex: /'/g, switch_state: "string_single" }
+		string_double: { regex: /"/g, switch_state: "string_double" }
 	},
 	// Single-line comments
 	comment_single: {
@@ -61,12 +60,6 @@ export default {
 		}
 	},
 	string_double: {
-		end: {
-			regex: /"/g,
-			switch_state: "default"
-		}
-	},
-	string_single: {
 		end: {
 			regex: /"/g,
 			switch_state: "default"