+++ src/langs/sh.mjs
@@ -42,7 +42,7 @@ export default {
 	},
 	string_single: {
 		end: {
-			regex: /"/g,
+			regex: /'/g,
 			switch_state: "default"
 		}
 	},