+++ tests/go/quotes-backtick-multiline.in.go
@@ -1,0 +1,7 @@
+const test_string = `yayaya
+yayay`
+
+
+func main() {
+	
+}