+++ tests/go/quotes-backtick.in.go
@@ -1,0 +1,6 @@
+const test_string = `yayayayayay`
+
+
+func main() {
+	
+}