+++ tests/go/quotes-nested.in.go
@@ -1,0 +1,8 @@
+const test_string = `sessions
+"yayayayay"
+an apple appears`
+
+
+func main() {
+	
+}