+++ tests/php/string-double-multiline.in.php
@@ -1,0 +1,8 @@
+<?php
+
+$a = "ora
+nge";
+
+function test_function() {
+	return 4;
+}