+++ tests/php/string-nested.in.php
@@ -1,0 +1,7 @@
+<?php
+
+$a = "'orange'";
+
+function test_function() {
+	return 4;
+}