+++ tests/rust/string-double-multiline.in.rs
@@ -1,0 +1,8 @@
+pub fn hello() {
+    print!("hello,
+world");
+}
+
+pub fn hello2() {
+    print!("hello, world");
+}