+++ tests/rust/string-double.in.rs
@@ -1,0 +1,7 @@
+pub fn hello() {
+    print!("hello, world");
+}
+
+pub fn hello2() {
+    print!("hello, world");
+}