+++ CHANGELOG.md
@@ -18,6 +18,7 @@ npm install --global skyliner # For using the CLI globally
 
 ## v0.3 (unreleased)
  - javascript: Fix single quotes
+ - go, php, rust, sh: fix single quote handling
  - Index line numbers from 0, not 1
  - Ensure that earlier Lexer sessions are not resumed after a new Lexer session begins
      - If you need to outline multiple things at once with `Skyliner.outline_iterate()`, it is advised you create multiple `Skyliner` class instances.